+++ tests/v2/01_schemas/test_schemas_dataset.py
@@ -27,6 +27,8 @@ async def test_schemas_dataset_v2():
         DatasetCreateV2(
             name="name", zarr_dir="/zarr", filters={"types": {"a": "b"}}
         )
+    # Test zarr_dir=None is valid
+    DatasetCreateV2(name="name", zarr_dir=None)
 
     dataset_create = DatasetCreateV2(
         name="name",
@@ -34,6 +36,9 @@ async def test_schemas_dataset_v2():
         zarr_dir="/tmp/",
     )
     assert dataset_create.zarr_dir == normalize_url(dataset_create.zarr_dir)
+
+    with pytest.raises(ValidationError):
+        DatasetImportV2(name="name", zarr_dir=None)
 
     dataset_import = DatasetImportV2(
         name="name",