+++ CHANGELOG.md
@@ -11,6 +11,8 @@
     * Now `pip install` uses `--no-cache` (\#1980).
 * API
     * Deprecate the `verbose` query parameter in `GET /api/v2/task/collect/{state_id}/` (\#1980).
+    * Add `project_dir` attribute to `UserSettings` (\#1990).
+    * Set a default for `DatasetV2.zarr_dir` (\#1990).
     * Combine the `args_schema_parallel` and `args_schema_non_parallel` query parameters in `GET /api/v2/task/` into a single parameter `args_schema` (\#1998).
 
 # 2.7.1